# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentStatusesResponse": {
        "properties": {
          "agent_outputs": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AgentOutputArtifact"
            },
            "title": "Agent Outputs",
            "type": "object"
          },
          "current_node": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Node"
          },
          "node_statuses": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Node Statuses",
            "type": "object"
          }
        },
        "required": [
          "node_statuses",
          "agent_outputs"
        ],
        "title": "AgentStatusesResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageResponse": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "role": {
            "enum": [
              "user",
              "assistant",
              "system",
              "tool"
            ],
            "title": "Role",
            "type": "string"
          },
          "type": {
            "default": "text",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "role",
          "content"
        ],
        "title": "MessageResponse",
        "type": "object"
      },
      "RequestSchema": {
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "resume_payload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterruptResumePayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "thread_id": {
            "title": "Thread Id",
            "type": "string"
          }
        },
        "required": [
          "thread_id"
        ],
        "title": "RequestSchema",
        "type": "object"
      },
      "StreamStartResponse": {
        "properties": {
          "status": {
            "enum": [
              "started",
              "running"
            ],
            "title": "Status",
            "type": "string"
          },
          "thread_id": {
            "title": "Thread Id",
            "type": "string"
          }
        },
        "required": [
          "status",
          "thread_id"
        ],
        "title": "StreamStartResponse",
        "type": "object"
      },
      "ThreadStateResponse": {
        "properties": {
          "agent_outputs": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AgentOutputArtifact"
            },
            "title": "Agent Outputs",
            "type": "object"
          },
          "interrupts": {
            "items": {
              "$ref": "#/components/schemas/HumanTickerSelection"
            },
            "title": "Interrupts",
            "type": "array"
          },
          "is_running": {
            "title": "Is Running",
            "type": "boolean"
          },
          "last_seq_id": {
            "title": "Last Seq Id",
            "type": "integer"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageResponse"
            },
            "title": "Messages",
            "type": "array"
          },
          "next": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next"
          },
          "node_statuses": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Node Statuses",
            "type": "object"
          },
          "resolved_ticker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved Ticker"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "thread_id": {
            "title": "Thread Id",
            "type": "string"
          }
        },
        "required": [
          "thread_id",
          "messages",
          "interrupts",
          "is_running",
          "node_statuses",
          "agent_outputs",
          "last_seq_id"
        ],
        "title": "ThreadStateResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Pure FastAPI implementation for FinGraph Valuation Agent (Path B)",
    "title": "Neuro-Symbolic Valuation Engine API",
    "version": "2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/artifacts/{artifact_id}": {
      "get": {
        "description": "Retrieve artifact data with HTTP caching.",
        "operationId": "get_artifact_api_artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifact"
      }
    },
    "/history/{thread_id}": {
      "get": {
        "description": "Retrieve persistent history with cursor pagination.",
        "operationId": "get_history_history__thread_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  },
                  "title": "Response Get History History  Thread Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get History"
      }
    },
    "/stream": {
      "post": {
        "description": "Start or resume a job.",
        "operationId": "stream_agent_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Agent"
      }
    },
    "/stream/{thread_id}": {
      "get": {
        "description": "Attach to an existing background job's event stream.",
        "operationId": "attach_stream_stream__thread_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Attach Stream"
      }
    },
    "/thread/{thread_id}": {
      "get": {
        "description": "Retrieve history and job status.",
        "operationId": "get_thread_history_thread__thread_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Thread History"
      }
    },
    "/thread/{thread_id}/agents": {
      "get": {
        "description": "Retrieve node statuses and financial reports for the dashboard.",
        "operationId": "get_agent_statuses_thread__thread_id__agents_get",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentStatusesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Agent Statuses"
      }
    }
  }
}